feat(weekChart): generate lines, improve style & refactor
