+++ src/components/Statistics/WeekChart/types.ts
@@ -1,0 +1,1 @@
+export type DataType = Record<string, string | number>;
